fix: clean code
